+++ lib/array.rb
@@ -1,6 +1,6 @@
 class Array
   include Philter::Base
-  def philter search=nil, options={}
+  def philter search=nil, options={}, &block
     options = {
         get:          nil,
         debug:        false
@@ -9,41 +9,32 @@ class Array
       puts philter_help if options[:debug]
       raise "Specify search parameter!"
     end
-    results = []
-    self.each do |item|
-      puts "item #{item.class.name} #{item}"                               if options[:debug]
-      if search.respond_to? :each
-        # search: {} or []
-        selected = phil_search_from_list search, item, options
-      elsif search.is_a?(Regexp)
-        # Search has one item
-        # search: 3 or search: 'a'
-        print " b.  "                                                      if options[:debug]
-        selected = phil_evaluate item, nil, search, :simple, options.merge(operator: '=~')
-        puts "  #{'=> X' if selected}"                                     if options[:debug]
+
+    apply_block = true
+    results =
+    case search.class.name
+      when "Hash"
+        print " h. "                                                      if options[:debug]
+        phil_search_by_attributes search, options
+      when "String"
+        print " s. "                                                      if options[:debug]
+        operator = phil_get_operator search
+        operator ? phil_search_with_operator(search, operator: operator) : self.grep(search)
+      when "Array"
+        print " a. "                                                      if options[:debug]
+        phil_search_by_array search, options
       else
-        # Search has one item
-        # search: 3 or search: 'a'
-        # search: '<3' or any other operator
-        print " c.  "                                                      if options[:debug]
-        selected = phil_evaluate item, nil, search, :simple, options.merge(operator: '==')
-        puts "  #{'=> X' if selected}"                                     if options[:debug]
-      end
-      if selected
-        results <<
-        if options[:get]
-          if item.respond_to? options[:get]
-            item.send options[:get]
-          elsif item.respond_to? '[]'
-            item[options[:get]]
-          end
-        else
-          item
-        end
-      end
+        puts " g. #{search}"                                              if options[:debug]
+        apply_block = false
+        self.grep(search, &block)
     end
+
     puts "------------------"                                                                     if options[:debug]
     puts " #{results ? results.size : 'No'} item#{results && results.size == 1 ? '' : 's'} found" if options[:debug]
-    results
+    if block_given? && apply_block
+      results.map{|item| block.call(item)}
+    else
+      results
+    end
   end
 end