+++ test/unit_test.rb
@@ -37,29 +37,48 @@ class UnitTest < Test::Unit::TestCase
 
   test 'philter fixnums with operators' do
     search    = '>2'
-    should_be = [3, 4, 5]
+    expected_res = [3, 4, 5]
     res = get_array_of_fixnums.philter search
-    assert res == should_be,                  "Searching #{search} in #{get_array_of_fixnums} result should be #{should_be}, not #{res}"
+    assert res == expected_res,               "Searching #{search} in #{get_array_of_fixnums} result should be #{expected_res}, not #{res}"
 
     search    = '< 3'
-    should_be = [1, 2]
+    expected_res = [1, 2]
     res = get_array_of_fixnums.philter search
-    assert res == should_be,                  "Searching #{search} in #{get_array_of_fixnums} result should be #{should_be}, not #{res}"
+    assert res == expected_res,               "Searching #{search} in #{get_array_of_fixnums} result should be #{expected_res}, not #{res}"
 
     search    = '<= 1'
-    should_be = [1]
+    expected_res = [1]
     res = get_array_of_fixnums.philter search
-    assert res == should_be,                  "Searching #{search} in #{get_array_of_fixnums} result should be #{should_be}, not #{res}"
+    assert res == expected_res,               "Searching #{search} in #{get_array_of_fixnums} result should be #{expected_res}, not #{res}"
 
     search    = '>= 2'
-    should_be = [2, 3, 4, 5]
+    expected_res = [2, 3, 4, 5]
     res = get_array_of_fixnums.philter search
-    assert res == should_be,                  "Searching #{search} in #{get_array_of_fixnums} result should be #{should_be}, not #{res}"
+    assert res == expected_res,               "Searching #{search} in #{get_array_of_fixnums} result should be #{expected_res}, not #{res}"
 
     search    = '!= 3'
-    should_be = [1, 2, 4, 5]
+    expected_res = [1, 2, 4, 5]
     res = get_array_of_fixnums.philter search
-    assert res == should_be,                  "Searching #{search} in #{get_array_of_fixnums} result should be #{should_be}, not #{res}"
+    assert res == expected_res,               "Searching #{search} in #{get_array_of_fixnums} result should be #{expected_res}, not #{res}"
+  end
+
+  test 'philter fixnums by a range' do
+    search = (2..4)
+    expected_res = [2, 3, 4]
+    res = get_array_of_fixnums.philter search
+    assert res == expected_res,               "Searching #{search} in #{get_array_of_fixnums} result should be #{expected_res}, not #{res}"
+  end
+
+  test 'philter fixnums with block' do
+    search = 2
+    expected_res = [(search*2)]
+    res = get_array_of_fixnums.philter(search){|n|n*2}
+    assert res == expected_res,               "Searching #{search} in #{get_array_of_fixnums} result should be #{expected_res}, not #{res}"
+
+    search = [1, 2]
+    expected_res = [2, 4]
+    res = get_array_of_fixnums.philter(search){|n|n*2}
+    assert res == expected_res,               "Searching #{search} in #{get_array_of_fixnums} result should be #{expected_res}, not #{res}"
   end
 
   #################
@@ -69,13 +88,13 @@ class UnitTest < Test::Unit::TestCase
   test 'philter strings' do
     search = 'a'
     res = get_array_of_strings.philter search
-    assert res == [search],                   "Searching #{search} in #{get_array_of_strings} result should be [#{search}]"
+    assert res == [search],                   "Searching #{search} in #{get_array_of_strings} result should be [#{search}], not #{res}"
   end
 
   test 'philter strings by an array of strings' do
     search = %w(a c)
     res = get_array_of_strings.philter search
-    assert res == search,                     "Searching #{search} in #{get_array_of_strings} result should be #{search}"
+    assert res == search,                     "Searching #{search} in #{get_array_of_strings} result should be #{search}, not #{res}"
   end
 
   #################
@@ -85,13 +104,13 @@ class UnitTest < Test::Unit::TestCase
   test 'philter symbols' do
     search = :a
     res = get_array_of_symbols.philter search
-    assert res == [search],                   "Searching #{search} in #{get_array_of_symbols} result should be [#{search}]"
+    assert res == [search],                   "Searching #{search} in #{get_array_of_symbols} result should be [#{search}], not #{res}"
   end
 
   test 'philter symbols by an array of symbols' do
     search = [:a, :c]
     res = get_array_of_symbols.philter search
-    assert res == search,                     "Searching #{search} in #{get_array_of_symbols} result should be #{search}"
+    assert res == search,                     "Searching #{search} in #{get_array_of_symbols} result should be #{search}, not #{res}"
   end
 
   #################
@@ -136,10 +155,6 @@ class UnitTest < Test::Unit::TestCase
     res = get_array_with_hashes.philter id: search
     assert res.size == search.size,           "Should return #{search.size} item, not #{res.size}"
 
-    bol_hash = true
-    res.each{|item| bol_hash = item.is_a?(Hash) unless bol_hash }
-    assert bol_hash,                          "Every item should be an Hash, not a #{res.first.class.name}"
-
     check_names = %w(Larry Bill)
     res.each do |item|
       assert check_names.include?(item[:name]), "Person #{item[:name]} should not be present, only #{check_names.join(', ')}"
@@ -174,10 +189,6 @@ class UnitTest < Test::Unit::TestCase
     res = get_array_with_hashes.philter name: search
     assert res.size == search.size,           "Should return #{search.size} item, not #{res.size}"
 
-    bol_hash = true
-    res.each{|item| bol_hash = item.is_a?(Hash) unless bol_hash }
-    assert bol_hash,                          "Every item should be an Hash, not a #{res.first.class.name}"
-
     check_ids = [3, 4]
     res.each do |item|
       assert check_ids.include?(item[:id]),   "Person id #{item[:id]} should not be present, only #{check_ids.join(', ')}"
@@ -188,11 +199,27 @@ class UnitTest < Test::Unit::TestCase
     res = get_array_with_hashes.philter email: /\A.+@fca/
     assert res.size == 3,                     "Should return three item, not #{res.size}"
 
-    bol_hash = true
-    res.each{|item| bol_hash = item.is_a?(Hash) unless bol_hash }
-    assert bol_hash,                          "Every item should be an Hash, not a #{res.first.class.name}"
-
     check_names = %w(Sergio Marco Luca)
+    res.each do |item|
+      assert check_names.include?(item[:name]), "Person #{item[:name]} should not be present, only #{check_names.join(', ')}"
+    end
+  end
+
+  test 'philter hashes by more attributes join by and' do
+    res = get_array_with_hashes.philter email: /\A.+@fca/, name: 'Sergio'
+    assert res.size == 1,                     "Should return 1 item, not #{res.size}"
+
+    check_names = %w(Sergio)
+    res.each do |item|
+      assert check_names.include?(item[:name]), "Person #{item[:name]} should not be present, only #{check_names.join(', ')}"
+    end
+  end
+
+  test 'philter hashes by more attributes join by or' do
+    res = get_array_with_hashes.philter({email: /\A.+@fca/, name: 'Larry'}, or: true)
+    assert res.size == 4,                     "Should return 4 item, not #{res.size}"
+
+    check_names = %w(Sergio Marco Luca Larry)
     res.each do |item|
       assert check_names.include?(item[:name]), "Person #{item[:name]} should not be present, only #{check_names.join(', ')}"
     end
@@ -319,6 +346,26 @@ class UnitTest < Test::Unit::TestCase
     end
   end
 
+  test 'philter objects by more attributes join by and' do
+    res = get_array_with_objects.philter email: /\A.+@fca/, name: 'Sergio'
+    assert res.size == 1,                     "Should return 1 item, not #{res.size}"
+
+    check_names = %w(Sergio)
+    res.each do |item|
+      assert check_names.include?(item.name), "Person #{item.name} should not be present, only #{check_names.join(', ')}"
+    end
+  end
+
+  test 'philter objects by more attributes join by or' do
+    res = get_array_with_objects.philter({email: /\A.+@fca/, name: 'Larry'}, or: true)
+    assert res.size == 4,                     "Should return 4 item, not #{res.size}"
+
+    check_names = %w(Sergio Marco Luca Larry)
+    res.each do |item|
+      assert check_names.include?(item.name), "Person #{item.name} should not be present, only #{check_names.join(', ')}"
+    end
+  end
+
   #################
   ### HYBRIDS
   #################
@@ -326,15 +373,15 @@ class UnitTest < Test::Unit::TestCase
   test 'philter hybrids' do
     search = 0
     res = get_array_of_hybrids.philter search
-    assert res == [search],                   "Searching #{search} in #{get_array_of_hybrids} result should be #{search}"
+    assert res == [search],                   "Searching #{search} in #{get_array_of_hybrids} result should be #{search}, not #{res}"
 
     search = 'a'
     res = get_array_of_hybrids.philter search
-    assert res == [search],                   "Searching #{search} in #{get_array_of_hybrids} result should be #{search}"
+    assert res == [search],                   "Searching #{search} in #{get_array_of_hybrids} result should be #{search}, not #{res}"
 
     search = :first
     res = get_array_of_hybrids.philter search
-    assert res == [search],                   "Searching #{search} in #{get_array_of_hybrids} result should be #{search}"
+    assert res == [search],                   "Searching #{search} in #{get_array_of_hybrids} result should be #{search}, not #{res}"
 
     res = get_array_of_hybrids.philter id: 1
     assert res.size         == 1,             "Should return one item"
@@ -345,6 +392,88 @@ class UnitTest < Test::Unit::TestCase
     assert res.size         == 1,             "Should return one item"
     assert res.first.is_a?(Employee), "The item is a Employee"
     assert res.first.tag    == 'third',       "Should return the hash with tag='third'"
+  end
+
+  test 'philter hybrids by class' do
+    search = Fixnum
+    expected_res = [0, 1, 2]
+    res = get_array_of_hybrids.philter search
+    assert res == expected_res,                "Searching #{search} in #{get_array_of_hybrids} result should be #{expected_res}, not #{res}"
+
+    search = String
+    expected_res = %w(a b c)
+    res = get_array_of_hybrids.philter search
+    assert res == expected_res,                "Searching #{search} in #{get_array_of_hybrids} result should be #{expected_res}, not #{res}"
+
+    search = Symbol
+    expected_res = [:first, :second, :third]
+    res = get_array_of_hybrids.philter search
+    assert res == expected_res,                "Searching #{search} in #{get_array_of_hybrids} result should be #{expected_res}, not #{res}"
+
+    search = Hash
+    expected_res = get_array_of_hybrids.select{|e| e.is_a? search}
+    res = get_array_of_hybrids.philter search
+    assert res == expected_res,                "Searching #{search} in #{get_array_of_hybrids} result should be #{expected_res}, not #{res}"
+
+    search = Employee
+    ar = get_array_of_hybrids
+    expected_res = ar.select{|e| e.is_a? search}
+    res = ar.philter search
+    assert res == expected_res,                "Searching #{search} in #{ar} result should be #{expected_res}, not #{res}"
+
+    search = Customer
+    ar = get_array_of_hybrids
+    expected_res = ar.select{|e| e.is_a? search}
+    res = ar.philter search
+    assert res == expected_res,                "Searching #{search} in #{ar} result should be #{expected_res}, not #{res}"
+
+    search = [Employee, Customer]
+    ar = get_array_of_hybrids
+    expected_res = ar.select{|e| search.include? e.class}
+    res = ar.philter search
+    assert res == expected_res,                "Searching #{search} in #{ar} result should be #{expected_res}, not #{res}"
+
+    search = {tag: Symbol}
+    ar = get_array_of_hybrids
+    # Include every object which has the id attribute
+    expected_res = ar.select do |e|
+      if e.respond_to?(:tag)
+        e.tag.is_a?(Symbol)
+      elsif e.is_a?(Hash) && e[:tag]
+        e[:tag].is_a?(Symbol)
+      end
+    end
+    res = ar.philter search
+    assert res == expected_res,                "Searching #{search} in #{ar} result should be #{expected_res}, not #{res}"
+
+    search = {tag: String}
+    ar = get_array_of_hybrids
+    # Include every object which has the id attribute
+    expected_res = ar.select do |e|
+      if e.respond_to?(:tag)
+        e.tag.is_a?(String)
+      elsif e.is_a?(Hash) && e[:tag]
+        e[:tag].is_a?(String)
+      end
+    end
+    res = ar.philter search
+    assert res == expected_res,                "Searching #{search} in #{ar} result should be #{expected_res}, not #{res}"
+  end
+
+  test 'philter hybrids by array of hybrids' do
+    search = [Employee, 1, 'a', :first]
+    ar = get_array_of_hybrids
+    expected_res = search
+    res = ar.philter search
+    assert res == expected_res,                "Searching #{search} in #{ar} result should be #{expected_res}, not #{res}"
+  end
+
+  test 'philter hybrids by symbol everywhere' do
+    search = {tag: :first}
+    ar = get_array_of_hybrids
+    expected_res = [:first, {:id=>1, :tag=>:first}]
+    res = ar.philter search, everywhere: true
+    assert res == expected_res,                "Searching #{search} in #{ar} result should be #{expected_res}, not #{res}"
   end
 
   test 'philter hybrids with an array of values' do