+++ CHANGELOG.md
@@ -1,4 +1,33 @@
-v0.7.0 [☰](https://github.com/marcomd/Philter/compare/v0.7.0...v0.6.0) May 24th, 2016
+v1.0.0 [☰](https://github.com/marcomd/Philter/compare/v0.8.0...v1.0.0) May 25th, 2016
+------------------------------
+* Code revolution: more simple, secure and readable
+* Performance: now it use grep for simple search
+* Test:
+    * wide coverage with 37 unit tests and 137 assertions
+    * updated 8 performance tests
+* Functional:
+    * now you can pass a block to update or completely change any array's element
+    * you can search multiple attributes with and boolean operator. It's also possible use `or` as option
+    * added option `everywhere` to search an attribute also in simple value.
+        Examples:
+        ```ruby
+        [:first, :second, {id: 1, tag: :first}, {id: 2, tag: :second}].philter tag: :first
+         => [{:id=>1, :tag=>:first}]
+        [:first, :second, {id: 1, tag: :first}, {id: 2, tag: :second}].philter({tag: :first}, everywhere: true)
+         => [:first, {:id=>1, :tag=>:first}]
+        ```
+
+
+v0.8.0 May 24th, 2016
+------------------------------
+* Added philter by class, example:
+    ```ruby
+    [:first, 'second'].philter Symbol
+     => [:first]
+    ```
+* Many internal changement
+
+v0.7.0 [☰](https://github.com/marcomd/Philter/compare/v0.6.0...v0.7.0) May 24th, 2016
 ------------------------------
 * Added performance tests